requirements.txt
accelerate~=0.21.0
bitsandbytes~=0.41.3
einops~=0.8.0
fastapi==0.83.0
fasttext-wheel~=0.9.2
huggingface_hub==0.23.5
hypercorn[h3]==0.14.4
hypercorn[trio]==0.14.4
hypercorn[uvloop]==0.14.4
numpy==1.22.0
pathspec==0.9.0
pydantic==1.9.2
python-multipart==0.0.5
sentence_transformers~=3.2.1
uvicorn==0.17.0